+++ src/methods/getPendingTransactions.ts
@@ -8,6 +8,7 @@ import {
   X_PLATFORM_APPID,
 } from '../constants';
 import { Item, PaymentOptions, Transaction, Transactions } from '../interfaces/marketplace';
+import { BuildPaymentOptions, BuildTradeItems } from '../utils/helperFunctions';
 import { ApiClient } from './apiClient';
 import { CheckToken } from './auth';
 
@@ -75,48 +76,4 @@ const MapResults = async (data: any): Promise<Transactions> => {
   });
 
   return transactions;
-};
-
-const BuildTradeItems = async (tradeItems: any): Promise<Item> => {
-  const defaultItem: Item = {
-    id: '',
-    assetUrl: '',
-    itemId: '',
-    name: '',
-    tags: [],
-    type: '',
-  };
-  if (!tradeItems || tradeItems.length === 0) {
-    return defaultItem;
-  }
-
-  const { id, itemId, assetUrl, name, tags, type } = tradeItems[0].item;
-
-  return {
-    id,
-    itemId,
-    assetUrl,
-    name,
-    tags,
-    type,
-  };
-};
-
-const BuildPaymentOptions = async (options: any): Promise<PaymentOptions> => {
-  const defualtPaymentOptions: PaymentOptions = {
-    quantity: 0,
-    price: 0,
-    transactionFee: 0,
-  };
-
-  if (!options || options.length === 0) return defualtPaymentOptions;
-
-  const { quantity } = options[0].item.viewer.meta;
-  const { price, transactionFee } = options[0];
-
-  return {
-    quantity,
-    price,
-    transactionFee,
-  };
 };
