+++ src/utils/helperFunctions.ts
@@ -19,6 +19,9 @@ import {
   ItemNamesDESC,
   ItemNamesASC,
   MarkplaceSearchType,
+  Item,
+  PaymentOptions,
+  Payment,
 } from '../interfaces/marketplace';
 
 // ============== Marketplace Helper Functions ==============
@@ -199,6 +202,66 @@ export const ReturnSortBy = async (
       if (searchType === 'sell') return PurchaseAvailableHighLow;
       else return SaleAvailableHighLow;
   }
+};
+
+export const BuildTradeItems = async (tradeItems: any): Promise<Item> => {
+  const defaultItem: Item = {
+    id: '',
+    assetUrl: '',
+    itemId: '',
+    name: '',
+    tags: [],
+    type: '',
+  };
+  if (!tradeItems || tradeItems.length === 0) {
+    return defaultItem;
+  }
+
+  const { id, itemId, assetUrl, name, tags, type } = tradeItems[0].item;
+
+  return {
+    id,
+    itemId,
+    assetUrl,
+    name,
+    tags,
+    type,
+  };
+};
+
+export const BuildPaymentOptions = async (options: any): Promise<PaymentOptions> => {
+  const defualtPaymentOptions: PaymentOptions = {
+    quantity: 0,
+    price: 0,
+    transactionFee: 0,
+  };
+
+  if (!options || options.length === 0) return defualtPaymentOptions;
+
+  const { quantity } = options[0].item.viewer.meta;
+  const { price, transactionFee } = options[0];
+
+  return {
+    quantity,
+    price,
+    transactionFee,
+  };
+};
+
+export const BuildPayment = async (payment: any): Promise<Payment> => {
+  const defualtPayment: Payment = {
+    price: 0,
+    transactionFee: 0,
+  };
+
+  if (!payment || payment.length === 0) return defualtPayment;
+
+  const { price, transactionFee } = payment;
+
+  return {
+    price,
+    transactionFee,
+  };
 };
 
 // ==========================================================
